+++ src/server/ccmap/var-link.ts
@@ -1,3 +1,4 @@
+import { InstanceinatorInstance } from 'cc-instanceinator/src/instance'
 import { prestart } from '../../loading-stages'
 
 declare global {
@@ -13,14 +14,17 @@ prestart(() => {
         clear() {
             this.parent()
             if (this.linkedTo) {
-                linkVars(this, this.linkedTo)
+                link(this, this.linkedTo)
             }
         },
     })
 })
-
-export function linkVars(to: ig.Vars, from: ig.Vars) {
+function link(to: ig.Vars, from: ig.Vars) {
     to.linkedTo = from
     to.storage.maps = from.storage.maps
     to.storage.plot = from.storage.plot ??= {}
+}
+
+export function linkVars(toInst: InstanceinatorInstance, fromInst: InstanceinatorInstance) {
+    link(toInst.ig.vars, fromInst.ig.vars)
 }