+++ src/server/music.ts
@@ -1,0 +1,5 @@
+import { InstanceinatorInstance } from 'cc-instanceinator/src/instance'
+
+export function linkMusic(to: InstanceinatorInstance, from: InstanceinatorInstance) {
+    to.ig.music = from.ig.music
+}