+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 
 ### Fixed
 
-- Link base instance options to server and map options
+- Fix some options weirdness by linking base instance sc.OptionModel to server and map sc.OptionModel
+- Fix some music weirdness by linking base instance ig.Music to server and map ig.Music
 
 ## [0.5.3] 2025-10-29
 ## [0.5.2] 2025-10-29